rename pgscatalog-combine to pgscatalog-format
